requirements.txt
Babel==2.6.0
Django==2.1.7
django-bootstrap-form==3.4
django-phonenumber-field==2.3.1
django-widget-tweaks==1.4.3
phonenumbers==8.10.10
pytz==2018.9
